+++ test_suite/tests/socket_multi_test.gd
@@ -1,6 +1,9 @@
 extends "res://base_test.gd"
 
 var content = {"My": "message"}
+var match_props = {"region": "europe"}
+var got_msg = false
+var got_match = false
 
 func setup():
 	var client = Nakama.create_client(Config.SERVER_KEY, Config.HOST, Config.PORT, Config.SCHEME)
@@ -15,6 +18,7 @@ func setup():
 
 	var socket1 = Nakama.create_socket_from(client)
 	socket1.connect("received_channel_message", self, "_on_socket1_message")
+	socket1.connect("received_matchmaker_matched", self, "_on_socket1_matchmaker_matched")
 	var done = yield(socket1.connect_async(session1), "completed")
 	# Check that connection succeded
 	if assert_false(done.is_exception()):
@@ -39,10 +43,30 @@ func setup():
 	if assert_false(msg_ack.is_exception()):
 		return
 
+	var ticket1 = yield(socket1.add_matchmaker_async("+properties.region:europe", 2, 8, match_props), "completed")
+	if assert_false(ticket1.is_exception()):
+		return
+	var ticket2 = yield(socket2.add_matchmaker_async("+properties.region:europe", 2, 8, match_props), "completed")
+	if assert_false(ticket2.is_exception()):
+		return
+
 func _on_socket1_message(msg):
 	if assert_equal(msg.content, JSON.print(content)):
 		return
-	done()
+	got_msg = true
+	check_end()
+
+func _on_socket1_matchmaker_matched(p_matchmaker_matched):
+	if assert_equal(p_matchmaker_matched.users[0].string_properties, JSON.print(match_props)):
+		return
+	if assert_equal(p_matchmaker_matched.users[1].string_properties, JSON.print(match_props)):
+		return
+	got_match = true
+	check_end()
 
 func _process(_delta):
 	assert_time(3)
+
+func check_end():
+	if got_match and got_msg:
+		done()