+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [keep a changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+- Fix Dictionary serialization (e.g. "NakamaSocket.add_matchmaker_async" "p_string_props").
 
 ## [2.1.0] - 2020-08-01